% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{common_samples}
\alias{common_samples}
\title{A matrix of common samples between tissues in sample_by_feature_data
Calculates the total non-NA expressions for a pair of tissues (of 53 tissues)
and records that in the cell corresponding to each tissue pair}
\format{A matrix of integers specifying number of common persons contributing a
pair of tissues for 53 tissues. The matrix is symmetric with 53 rows and 53 columns.}
\usage{
common_samples
}
\description{
A matrix of common samples between tissues in sample_by_feature_data
Calculates the total non-NA expressions for a pair of tissues (of 53 tissues)
and records that in the cell corresponding to each tissue pair
}
\keyword{datasets}
