% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorShrink2Data.R
\name{CorShrink2Data}
\alias{CorShrink2Data}
\title{Box-Shrinkage of correlations from a data matrix with missing data}
\usage{
CorShrink2Data(data_with_missing)
}
\arguments{
\item{data_with_missing}{The samples by features data matrix. May contain NA values.}
}
\description{
Performs a box shrinkage of correlations from a data matrix with missing data
using the Fisher Z-score formulation
}
\examples{
data("sample_by_feature_data")
out = CorShrink2Data(sample_by_feature_data)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
