% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorShrink2DataLoss.R
\name{CorShrink2DataLoss}
\alias{CorShrink2DataLoss}
\title{CorShrink Loss method for correlation shrinkage}
\usage{
CorShrink2DataLoss(data_with_missing, alpha = 1)
}
\arguments{
\item{data_with_missing}{The samples by features data matrix. May contain NA values.}

\item{alpha}{The tuning parameter for the L-1 scaling.}
}
\description{
Performs the CorShrink-Loss approach of correlation shrinkage, that
             uses the Taylor series expansion of the empirical correlation matrix
             around the population correlation matrix to derive the distribution.
}
\examples{
data("sample_by_feature_data")
out = CorShrink2DataLoss(sample_by_feature_data, alpha = 1)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")
out2 = CorShrink2DataLoss(sample_by_feature_data, alpha = 10)
corrplot::corrplot(out2, diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")
}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
