% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorShrink2DataSlack.R
\name{CorShrink2DataSlack}
\alias{CorShrink2DataSlack}
\title{Box-Shrinkage of correlations from a data matrix with missing data using CorShrink-sparse with slack variables}
\usage{
CorShrink2DataSlack(data_with_missing, alpha = 1)
}
\arguments{
\item{data_with_missing}{The samples by features data matrix. May contain NA values.}

\item{alpha}{The tuning parameter for the L-1 shrinkage of the slack variables}
}
\description{
Performs a box shrinkage of correlations from a data matrix with missing data
using the Fisher Z-score formulation, similar to CorShrink-sparse but slacking the
box constraint using slack variables.
}
\examples{
data("sample_by_feature_data")
out = CorShrink2DataSlack(sample_by_feature_data, alpha = 1)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
