% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorShrink-word2vec.R
\name{CorShrink_word2vec}
\alias{CorShrink_word2vec}
\title{CorShrink on word rankings in word2vec model}
\usage{
CorShrink_word2vec(model_true, model_boot_list, word_vec,
  num_related_words = 50, ash.control = list())
}
\arguments{
\item{model_true}{the true word2vec model}

\item{model_boot_list}{a list of bootstrap word2vec models with resampling lines of text.}

\item{word_vec}{A vector of words whose rankings are to be modified}

\item{num_related_words}{Number of related words to consider per word in \code{word_vec}}

\item{ash.control}{The control parameters for adaptive shrinkage}
}
\value{
Returns adaptively shrunk cosine similarity values of the words of interest in \code{word_vec}.
}
\description{
This function computes word rankings on the word2vec model from CorShrink
             using resampling methods.
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}
