% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorShrinkML.R
\name{CorShrinkML}
\alias{CorShrinkML}
\title{Adaptive shrinkage of a correlation matrix using ML estimation}
\usage{
CorShrinkML(cormat, nsamp_mat, sd_boot = FALSE, cor_transform_sd_vec = NULL,
  thresh_up = 0.99, thresh_down = -0.99, image = FALSE, tol = 1e-06,
  ash.control = list())
}
\arguments{
\item{cormat}{The sample correlation matrix}

\item{nsamp_mat}{The number of samples/ matrix of samples over which the correlation matrix is estimated.}

\item{sd_boot}{Boolean variable indicating if bootstrap sd to be used or not.}

\item{cor_transform_sd_vec}{If \code{sd_boot} is not NULL, the bootstrap standard errors
need to be provided. These standard errors can be obtained from \code{bootcorSE_calc()} function.}

\item{image}{if TRUE, plots an image of the shrunk and non-shrunk correlation matrices}

\item{tol}{The tolerance to check the difference between ash-cor only and ash-cor PD matrices.}

\item{ash.control}{The control parameters for adaptive shrinkage}
}
\value{
Returns a shrunk version of the sample correlation matrix
        (the output is also a correlation matrix)
}
\description{
This function performs adaptive shrinkage of a sample correlation matrix using a
mixture normal prior on Fisher z-scores with fixed grid variances and the MLEs of
mixture proportions are calculated using the EM algorithm.
}
\references{
False Discovery Rates: A New Deal. Matthew Stephens bioRxiv 038216; doi: http://dx.doi.org/10.1101/038216
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}
