% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorShrinkMLvec.R
\name{CorShrinkMLvec}
\alias{CorShrinkMLvec}
\title{Adaptive shrinkage of a vector of correlations using ML estimation}
\usage{
CorShrinkMLvec(corvec, nsamp_vec, sd_boot = FALSE,
  cor_transform_sd_vec = NULL, thresh_up = 0.99, thresh_down = -0.99,
  ash.control = list())
}
\arguments{
\item{corvec}{The vector of sample correlations}

\item{nsamp_vec}{vector of the number of samples over which the correlation matrix is estimated.}

\item{sd_boot}{Boolean variable indicating if bootstrap sd to be used or not.}

\item{cor_transform_sd_vec}{If \code{sd_boot} is not NULL, the vector of bootstrap standard errors
need to be provided. These standard errors can be obtained from \code{bootcorSE_calc()} function.}

\item{thresh_up}{The upper threshold for correlations}

\item{thresh_down}{The lower threshold for correlations}

\item{ash.control}{The control parameters for adaptive shrinkage}
}
\value{
Returns an adaptively shrunk version of the vector of correlations.
}
\description{
This function performs adaptive shrinkage of a sample correlation matrix using a
mixture normal prior on Fisher z-scores with fixed grid variances and the MLEs of
mixture proportions are calculated using the EM algorithm.
}
\references{
False Discovery Rates: A New Deal. Matthew Stephens bioRxiv 038216; doi: http://dx.doi.org/10.1101/038216
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}
