% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorShrinkVEM2.R
\name{CorShrinkVEM2}
\alias{CorShrinkVEM2}
\title{Adaptive shrinkage of a correlation matrix using Variational EM on both mixture proportions and grid variances.}
\usage{
CorShrinkVEM2(cormat, nsamples, image = FALSE, tol = 1e-06,
  nullweight = 10, null.comp = 1)
}
\arguments{
\item{cormat}{The estimated sample correlation matrix}

\item{nsamples}{The number of samples over which the correlation matrix is estimated.}

\item{image}{if TRUE, plots an image of the shrunk and non-shrunk correlation matrices}

\item{tol}{The tolerance to check the difference between ash-cor only and ash-cor PD matrices.}

\item{nullweight}{The weight of the null component of the ash prior.}

\item{null.comp}{The number of null components. Default is 1.}
}
\value{
Returns a shrunk version of the sample correlation matrix (the output is also a correlation matrix)
}
\description{
This function performs adaptive shrinkage of a correlation matrix with a
mixture normal prior on Fisher z-scores, where the variances of the components are Inverse
Gamma distributed and mixture proportions are Dirichlet distributed with high weights on the
null component. The model is estimated using Variational EM.
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}
