% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovShrink.R
\name{CovShrink}
\alias{CovShrink}
\title{Adaptive shrinkage of a covariance matrix through shrinking of correlations.}
\usage{
CovShrink(data, nsamples, sd_boot = FALSE, partial = FALSE, type = c("ML",
  "VEM"), nboot = 50, nullweight = 10, null.comp = 1, thresh_up = 0.999,
  thresh_down = 0.001, tol = 1e-06, ash.control = list())
}
\arguments{
\item{data}{The samples by features data matrix.}

\item{nsamples}{The number of samples used for estimating correlation matrix.}

\item{type}{The type of adaptive correlation matrix shrinkage to be used.}

\item{nullweight}{The weight on the null component for the mixture normal prior used. Defaults to 10.}

\item{null.comp}{The number of null components chosen. Defaults to 1.}

\item{ash.control}{The control parameters for adaptive shrinkage}
}
\value{
Returns a shrunk version of the sample covariance matrix (the output is also a covariance matrix)
}
\description{
Performs adaptive shrinkage of covariance matrix through shrinkage of the
underlying correlation matrix using wither of the three schemes \code{CorShrinkML},
\code{CorShrinkVEM} or \code{CorShrinkVEM2}.
}
\references{
False Discovery Rates: A New Deal. Matthew Stephens bioRxiv 038216; doi: http://dx.doi.org/10.1101/038216
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}
