% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cov.R
\name{generate_cov}
\alias{generate_cov}
\title{Generate a covariance matrix}
\usage{
generate_cov(dim)
}
\arguments{
\item{dim}{The dimensionality of the covariance matrix to generate}
}
\value{
Returns a covariance matrix of a given dimensionality along with its eigenvalues vector.
}
\description{
Wrapper for genPositiveDefMat function in clusterGeneration package to generate covariance matrix
}
\keyword{covariance}
\keyword{eigenvalues}
\keyword{matrix,}

