% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCorShrink2Data.R
\name{iCorShrink2Data}
\alias{iCorShrink2Data}
\title{Adaptive regularized GLASSO of partial correlations from a data matrix with missing data}
\usage{
iCorShrink2Data(data_with_missing, alpha, expo = 0.05, shift = 0.01,
  lambda = 0.8, max_iter = 10, epsilon = 0.001)
}
\arguments{
\item{data_with_missing}{The samples by features data matrix. May contain NA values.}

\item{alpha}{The tuning parameter}

\item{expo}{The exponent on the scaling used in the adaptive regularization of tuning parameter}

\item{shift}{The shift in the scaling used in adaptive regularization of tuning parameter}

\item{lambda}{The weight on the constraint for sample size bias}

\item{max_iter}{The maximum number of iterations for the adaptive GLASSO run.}

\item{epsilon}{The tolerance level for the relative error specifying when to stop}
}
\description{
Performs an adaptove regularized GLASSO of partial correlations from a data matrix with missing data
using the Fisher Z-score formulation
}
\examples{
data("sample_by_feature_data")
out = iCorShrink2Data(sample_by_feature_data, alpha = 0.1)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
