% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCorShrinkData.R
\name{iCorShrinkData}
\alias{iCorShrinkData}
\title{Box-Shrinkage of partial correlations from a data matrix with missing data}
\usage{
iCorShrinkData(data_with_missing, alpha, lambda = 1)
}
\arguments{
\item{data_with_missing}{The samples by features data matrix. May contain NA values.}

\item{alpha}{The shrinkage intensity}

\item{lambda}{The weighting of the constraint in the shrinkage. Default lambda taken to be 1.}
}
\description{
Performs a box shrinkage of partial correlations from a data matrix with missing data
using the Fisher Z-score formulation
}
\examples{
data("sample_by_feature_data")
out = iCorShrinkData(sample_by_feature_data, alpha = 0.1)
corrplot::corrplot(as.matrix(out), diag = FALSE,
        col = colorRampPalette(c("blue", "white", "red"))(200),
        tl.pos = "td", tl.cex = 0.4, tl.col = "black",
        rect.col = "white",na.label.col = "white",
        method = "color", type = "upper")

}
\keyword{box}
\keyword{correlation}
\keyword{shrinkage,}
