% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairwise_corr_matrix}
\alias{pairwise_corr_matrix}
\title{An example matrix of pairwise correlations
A matrix of pairwise correlations computed from the sample_by_feature_data.}
\format{A  symmetric matrix of pairwise correlations with 53 rows and 53 columns.
Each entry is a pairwise correlations from the 53 features (tissues) in
sample_by_feature_data.}
\usage{
pairwise_corr_matrix
}
\description{
An example matrix of pairwise correlations
A matrix of pairwise correlations computed from the sample_by_feature_data.
}
\keyword{datasets}
