% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectDataCorShrink.R
\name{ProjectDataCorShrink}
\alias{ProjectDataCorShrink}
\title{Project data on the shrunk correlation or the covariance model space.}
\usage{
ProjectDataCorShrink(data, nsamples, nullweight = 10, null.comp = 1,
  type = c("ML", "VEM", "VEM2"))
}
\arguments{
\item{data}{The samples by features data matrix.}

\item{nsamples}{The number of samples used for estimating correlation matrix.}

\item{nullweight}{The weight on the null component for the mixture normal prior used. Defaults to 10.}

\item{null.comp}{The number of null components chosen. Defaults to 1.}

\item{type}{The type of adaptive correlation matrix shrinkage to be used.}
}
\value{
Returns a shrunk version of the sample covariance matrix (the output is also a covariance matrix)
}
\description{
Performs projection of the sample data into the the shrunk
correlation or the covariance model space, where the shrinkage is performed adaptively
on the covariance model using the \code{CovShrink} function which in turn uses one of the
three correlation shrinkage methods - \code{CorShrinkML},
\code{CorShrinkVEM} or \code{CorShrinkVEM2}.
}
\references{
False Discovery Rates: A New Deal. Matthew Stephens bioRxiv 038216; doi: http://dx.doi.org/10.1101/038216
}
\keyword{adaptive}
\keyword{correlation}
\keyword{shrinkage,}
