% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_by_feature_data}
\alias{sample_by_feature_data}
\title{Samples by features data matrix with missing values
A matrix with 544 samples (persons) and 53 features (tissues) of GTEx expression data
for gene ENSG00000166819 (https://www.gtexportal.org/home/).}
\format{A matrix with 544 rows and 53 columns, containing missing values coded as NA}
\usage{
sample_by_feature_data
}
\description{
Samples by features data matrix with missing values
A matrix with 544 samples (persons) and 53 features (tissues) of GTEx expression data
for gene ENSG00000166819 (https://www.gtexportal.org/home/).
}
\keyword{datasets}
